Refactor importer (#5)
* Refactor file parsers, but causes issues with order of execution

* Oops, I had not properly named these classes :-)

* Test new, refactored zip file parser class
