+++ spec/support/gtfs_import_helpers.rb
@@ -1,3 +1,5 @@
+require 'zip'
+
 module GtfsImportHelpers
   def headers
     {
